requirements.txt
requests>=2.31.0
yt-dlp>=2023.12.30
deep-translator>=1.11.4
youtube-transcript-api>=0.6.1
python-dotenv>=1.0.0
PyPDF2>=3.0.1
python-docx>=1.1.0
